+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   * Better quote support: support for `“` and missing end quote support (by Zetrypio).
   * Allow switching to the next variable(s) with tab key after autocompletion.
 * Added an Error Dialog, that will allow user to see what caused a crash, and copy the information to Clipboard if they wish.
+* Missing editor and color themes no longer crash the IDE and make it refuse to open until the theme is back somehow, instead it just switches back to defaults.
 
 ## Version 1.4.2 - 31.05.2025